move lexer out of racc file
